Support figures in R snippets
This involves factoring out the logic which sets output directories into
a common helper.  That way, both the `Rules` and the helper functions
can call this logic from a common source.

The key insight is to set the base.dir option to the desired output
directory.

We also set the fig.path to empty, but this is just a matter of
preference.