LoopingCholesky() function
For Cholesky of compact support covariance matrix which we intend to run
continuously, the only row that matters is the last.  Conceptually, at
each timestep, we shift the independent random samples by one position,
and replace the lost value with a new sample.  But it's computationally
more efficient to replace one value *without* shifting, and just take
the dot product with a shifted Cholesky row (since we can precompute all
possibilities).

This matrix is the set of all such precomputed rows.