Force the "object constancy" data to stay synced
In practice, it was synced up already, but this should guarantee it.

I also took the opportunity to refactor, and move the IDs from the
dataset to the plot specs.  This makes more sense intrinsically, because
the data is identical (which is the whole point of the section!), and
the IDs are purely a function of how we choose to do the plotting.