Add Dockerfile with all dependencies
I can't generate my site without installing a bunch of dependencies.  I
blog so rarely that every time I try to, _something_ has broken.

The fix is to get all my dependencies explicitly listed in a Dockerfile.

More work remains: it's currently creating files as `root`.  Also, the
interfaces are not so good: I need to make a script.  But this is a good
checkpoint!